// src/enums/PayrollIdentifier.ts
/* eslint-disable no-unused-vars */
export enum PayrollIdentifier {
  PERSAL = '0001', // All Government employees except SANDF
  SANDF = '0002', // South African National Defence Force
  SARS = '0004', // South African Revenue Services
  TELKOM = '0005', // Telkom SA
  SAPO = '0008', // South African Post Office
  TSHWANE = '0011', // TSHWANE Metropolitan Council
  ESKOM = '0015', // Eskom Holdings
  FORD = '0019', // Ford Motor Company of Southern Africa
  MANGAUNG = '0020', // Mangaung Local Municipality
  SASSA = '0041', // South African Social Security Agency
  ESKOM_PENSIONERS = '0042', // Eskom Pensioners Fund
  TRANSNET = '0044', // Transnet
  RTMC = '0045', // Road Traffic Management Corporation
  NELSON_MANDELA_BAY = '0074', // Nelson Mandela Bay Municipality
  SOCPEN = '0087', // SASSA Social Grants
  PRASA = '0088', // PRASA
  SAA = '0089', // South African Airways
  BMA = '0094', // Border Management Authority
  PEP = '0095' // PEP
}