// src/enums/TransactionType.ts
/* eslint-disable no-unused-vars */
export enum TransactionType {
  EMPLOYEE_ENQUIRIES = '1',
  Q_LINK_TRANSACTIONS = '5',
  COMMUNICATION_TEST = '6',
  REQUEST_NEW_PASSWORD = '7',
  REQUEST_NEXT_UPDATE_SEQ = '8',
  MAINTENANCE_FUEL_TRANSACTIONS = '10',
  MAINTENANCE_TRANSACTIONS = '11'
}
// Note: Other definitions will be added when the requirement
// is defined. TRX values 2, 3 and 4 are reserved for
// consolidation of loans, which has been de- activated at present